requirements/package.txt
six

django-modeltranslation
easy-thumbnails

django-object-actions
django-filter
tablib

